adicionando escrita de aequivo na main

+++ include/Registro.hpp
@@ -5,12 +5,12 @@ class Registro{
     private:
         int chave;
         char ** cadeia;
-        double *numerosReais;
+        float *numerosReais;
     
     public:
         int getChave(){ return chave;}
         void setChave(int chave){ this->chave = chave;}
-        void setRegistro(int chave, char ** cadeia, double * numerosReais);
+        void setRegistro(int chave, char ** cadeia, float * numerosReais);
 };
 
 #endif
