+++ src/Registro.cpp
@@ -1,6 +1,6 @@
 #include "Registro.hpp"
 
-void Registro::setRegistro(int chave, char ** cadeia, double * numerosReais){
+void Registro::setRegistro(int chave, char ** cadeia, float * numerosReais){
     this->chave = chave;
     this->cadeia = cadeia;
     this->numerosReais = numerosReais;
